smaller approach to dev check
