+++ .mocharc.yml
@@ -5,3 +5,5 @@ extension:
   - ts
 node-option:
   - 'loader=ts-node/esm/transpile-only'
+require:
+  - resources/mocha-loader.js
